requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
reportlab==4.2.2
sqlparse==0.5.1
tzdata==2024.1
